Update fastcgi_pass host value
The fastcgi_pass host value need to be manually changed. An error occurs :the ip of the upstream seems to magically switch when multiple docker-compose projects at running at the same time, with the same services name such as 'php'. This issue requires an investigation (it might be a docker-compose or docker issue)